+++ Plugins/Creature/Creature.hpp
@@ -133,6 +133,7 @@ private:
     ArgumentStack GetNoPermanentDeath           (ArgumentStack&& args);
     ArgumentStack SetNoPermanentDeath           (ArgumentStack&& args);
     ArgumentStack ComputeSafeLocation           (ArgumentStack&& args);
+    ArgumentStack DoPerceptionUpdateOnCreature  (ArgumentStack&& args);
 
     CNWSCreature *creature(ArgumentStack& args);
     std::unordered_map<uint8_t, std::unordered_map<ObjectID, int16_t>> m_RollModifier;